+++ src/app/api/chat/[provider]/agentRuntime.ts
@@ -22,6 +22,7 @@ import {
   LobePerplexityAI,
   LobeRuntimeAI,
   LobeTogetherAI,
+  LobeZeroOneAI,
   LobeZhipuAI,
   ModelProvider,
 } from '@/libs/agent-runtime';
@@ -185,6 +186,11 @@ class AgentRuntime {
         runtimeModel = this.initTogetherAI(payload);
         break;
       }
+
+      case ModelProvider.ZeroOne: {
+        runtimeModel = this.initZeroOne(payload);
+        break;
+      }
     }
 
     return new AgentRuntime(runtimeModel);
@@ -306,6 +312,13 @@ class AgentRuntime {
     const apiKey = apiKeyManager.pick(payload?.apiKey || TOGETHERAI_API_KEY);
 
     return new LobeTogetherAI({ apiKey });
+  }
+
+  private static initZeroOne(payload: JWTPayload) {
+    const { ZEROONE_API_KEY } = getServerConfig();
+    const apiKey = apiKeyManager.pick(payload?.apiKey || ZEROONE_API_KEY);
+
+    return new LobeZeroOneAI({ apiKey });
   }
 }