+++ src/app/api/errorResponse.ts
@@ -56,8 +56,11 @@ const getStatus = (errorType: ILobeAgentRuntimeErrorType | ErrorType) => {
     case AgentRuntimeErrorType.GroqBizError: {
       return 482;
     }
-    case AgentRuntimeErrorType.TogetherAIBizError: {
+    case AgentRuntimeErrorType.ZeroOneBizError: {
       return 483;
+    }
+    case AgentRuntimeErrorType.TogetherAIBizError: {
+      return 484;
     }
   }
   return errorType as number;