+++ src/config/modelProviders/index.ts
@@ -11,6 +11,7 @@ import OpenAIProvider from './openai';
 import OpenRouterProvider from './openrouter';
 import PerplexityProvider from './perplexity';
 import TogetherAIProvider from './togetherai';
+import ZeroOneProvider from './zeroone';
 import ZhiPuProvider from './zhipu';
 
 export const LOBE_DEFAULT_MODEL_LIST: ChatModelCard[] = [
@@ -26,6 +27,7 @@ export const LOBE_DEFAULT_MODEL_LIST: ChatModelCard[] = [
   TogetherAIProvider.chatModels,
   PerplexityProvider.chatModels,
   AnthropicProvider.chatModels,
+  ZeroOneProvider.chatModels,
 ].flat();
 
 export { default as AnthropicProvider } from './anthropic';
@@ -39,4 +41,5 @@ export { default as OpenAIProvider } from './openai';
 export { default as OpenRouterProvider } from './openrouter';
 export { default as PerplexityProvider } from './perplexity';
 export { default as TogetherAIProvider } from './togetherai';
+export { default as ZeroOneProvider } from './zeroone';
 export { default as ZhiPuProvider } from './zhipu';