+++ src/features/AgentSetting/AgentConfig/index.tsx
@@ -58,6 +58,7 @@ const AgentConfig = memo(() => {
                   value: 'docs',
                 },
               ]}
+              unoptimized={false}
               value={displayMode}
               width={144}
             />