+++ src/features/Conversation/Error/APIKeyForm/index.tsx
@@ -16,6 +16,7 @@ import OpenAIForm from './OpenAI';
 import OpenRouterForm from './OpenRouter';
 import PerplexityForm from './Perplexity';
 import TogetherAIForm from './TogetherAI';
+import ZeroOneForm from './ZeroOne';
 import ZhipuForm from './Zhipu';
 
 interface APIKeyFormProps {
@@ -68,6 +69,10 @@ const APIKeyForm = memo<APIKeyFormProps>(({ id, provider }) => {
 
       case ModelProvider.TogetherAI: {
         return <TogetherAIForm />;
+      }
+
+      case ModelProvider.ZeroOne: {
+        return <ZeroOneForm />;
       }
 
       default: