+++ src/services/_auth.ts
@@ -74,6 +74,10 @@ export const getProviderAuthPayload = (provider: string) => {
       return { apiKey: modelProviderSelectors.togetheraiAPIKey(useGlobalStore.getState()) };
     }
 
+    case ModelProvider.ZeroOne: {
+      return { apiKey: modelProviderSelectors.zerooneAPIKey(useGlobalStore.getState()) };
+    }
+
     default:
     case ModelProvider.OpenAI: {
       const openai = modelProviderSelectors.openAIConfig(useGlobalStore.getState());